@@ -16,7 +16,7 @@
         "source": "https://gitlab.com/datetimetools/datetimeinterval"
     },
     "require": {
-        "php": "^7.3"
+        "php": "^7.4"
     },
     "require-dev": {
         "php-parallel-lint/php-parallel-lint": "^1",
